% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{ADJUSTED_SALARY}
\alias{ADJUSTED_SALARY}
\title{ADJUSTED_SALARY}
\usage{
ADJUSTED_SALARY(c_jobtitle_hr, c_jobtitle_yr)
}
\arguments{
\item{c_jobtitle_hr}{hourly}

\item{c_jobtitle_yr}{yearly}
}
\description{
per hour
take into account pension and
national insurance contributions
}
