% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{CSITE_SCREEN_PHLEB}
\alias{CSITE_SCREEN_PHLEB}
\title{CSITE_SCREEN_PHLEB}
\usage{
CSITE_SCREEN_PHLEB(n_id, n_screen)
}
\arguments{
\item{n_id}{number identified}

\item{n_screen}{}
}
\description{
CSITE_SCREEN_PHLEB
}
