% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_parameters}
\alias{load_parameters}
\alias{load_parameter}
\title{Load parameters from the central CSV file}
\usage{
load_parameters(envir)

load_parameter(envir)
}
\arguments{
\item{envir}{The environment to load the parameters into}
}
\description{
Load parameters from the central CSV file
}
