% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{PATH_INVITE_BIRM}
\alias{PATH_INVITE_BIRM}
\title{PATH_INVITE_BIRM}
\usage{
PATH_INVITE_BIRM(n_id, n_screen, n_latent)
}
\arguments{
\item{n_id}{number identified}

\item{n_screen}{}

\item{n_latent}{}
}
\description{
PATH_INVITE_BIRM
}
