% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{PATH_SITE_BIRM}
\alias{PATH_SITE_BIRM}
\title{PATH_SITE_BIRM}
\usage{
PATH_SITE_BIRM(n_id, n_screen, n_latent, phleb_thresh = 25)
}
\arguments{
\item{n_id}{number identified}

\item{n_screen}{}

\item{phleb_thresh}{numbered screened to use external company for taking blood}
}
\description{
PATH_SITE_BIRM
}
