% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Package environment to store costing parameters}
\format{
An object of class \code{environment} of length 54.
}
\usage{
pkg_env
}
\description{
Package environment to store costing parameters
}
\keyword{datasets}
