% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{total_year_cost}
\alias{total_year_cost}
\title{Total cost per year}
\usage{
total_year_cost(inc_sample, id_per_inc, screen_per_inc, ltbi_per_inc)
}
\arguments{
\item{inc_sample}{number of incidents}

\item{id_per_inc}{number identified per incident}

\item{screen_per_inc}{number screened per incident}

\item{ltbi_per_inc}{number of ltbi per incident}
}
\value{
real
}
\description{
this is the main overall cost function.
}
\examples{

total_year_cost(10,10,5,1)

}
