% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBA_converted.R
\name{vtotal_year_cost}
\alias{vtotal_year_cost}
\title{Vectorised costs
in order to use in mutate()}
\usage{
vtotal_year_cost(inc_sample, id_per_inc, screen_per_inc, ltbi_per_inc)
}
\description{
Vectorised costs
in order to use in mutate()
}
